[e2e] Modify existing search test instead
